WIP: rename M3GNet/MEGNet to *_dgl.py + lift EmbeddingBlock
- Rename _m3gnet.py/_megnet.py to *_dgl.py (TensorNet pattern); leave
  thin shims at the original paths so checkpoints whose model.json
  records "@module": "matgl.models._m3gnet" still load.
- Lift EmbeddingBlock into _embedding.py (no DGL coupling) so both
  backends can share it; re-export the old import path from
  _embedding_dgl.py.
- Refactor ThreeBodyInteractions.forward and
  SphericalBesselWithHarmonics.forward to take tensor inputs so a
  PyG caller can drive them. Update the DGL caller and the
  three-body test accordingly. SphericalBesselWithHarmonics keeps
  a backward-compat path that accepts a DGL line graph.
