+++ src/matgl/layers/__init__.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 
 from matgl.config import BACKEND
 from matgl.layers._activations import ActivationFunction
+from matgl.layers._basis import FourierExpansion, RadialBesselFunction, SphericalBesselWithHarmonics
 from matgl.layers._bond import BondExpansion
 from matgl.layers._core import (
     MLP,
@@ -11,13 +12,14 @@ from matgl.layers._core import (
     GatedMLP,
     build_gated_equivariant_mlp,
 )
+from matgl.layers._embedding import EmbeddingBlock
 from matgl.layers._norm import GraphNorm
+from matgl.layers._three_body import ThreeBodyInteractions
 
 if BACKEND == "DGL":
     from matgl.layers._atom_ref_dgl import AtomRef
-    from matgl.layers._basis import FourierExpansion, RadialBesselFunction, SphericalBesselWithHarmonics
     from matgl.layers._core_dgl import EdgeSet2Set, GatedMLPNorm, MLPNorm
-    from matgl.layers._embedding_dgl import EmbeddingBlock, NeighborEmbedding, TensorEmbedding
+    from matgl.layers._embedding_dgl import NeighborEmbedding, TensorEmbedding
     from matgl.layers._graph_convolution_dgl import (
         CHGNetAtomGraphBlock,
         CHGNetBondGraphBlock,
@@ -38,7 +40,6 @@ if BACKEND == "DGL":
         WeightedReadOut,
         WeightedReadOutPair,
     )
-    from matgl.layers._three_body import ThreeBodyInteractions
     from matgl.layers._zbl_dgl import NuclearRepulsion
 else:
     from matgl.layers._atom_ref_pyg import AtomRefPyG  # type: ignore[assignment]
