+++ src/matgl/models/__init__.py
@@ -8,8 +8,8 @@ from ._core import MatGLModel
 
 if BACKEND == "DGL":
     from ._chgnet import CHGNet
-    from ._m3gnet import M3GNet
-    from ._megnet import MEGNet
+    from ._m3gnet_dgl import M3GNet
+    from ._megnet_dgl import MEGNet
     from ._qet_dgl import QET
     from ._so3net import SO3Net
     from ._tensornet_dgl import TensorNet
