+++ src/matgl/models/_megnet.py
@@ -1,224 +1,13 @@
-"""Implementation of MatErials Graph Network (MEGNet) model.
+"""Backward-compat shim: ``MEGNet`` moved to ``_megnet_dgl``.
 
-Graph networks are a new machine learning (ML) paradigm that supports both relational reasoning and combinatorial
-generalization. For more details on MEGNet, please refer to::
-
-    Chen, C.; Ye, W.; Zuo, Y.; Zheng, C.; Ong, S. P. _Graph Networks as a Universal Machine Learning Framework for
-    Molecules and Crystals._ Chem. Mater. 2019, 31 (9), 3564-3572. DOI: 10.1021/acs.chemmater.9b01294.
+Pretrained checkpoints saved with ``"@module": "matgl.models._megnet"`` rely on
+this module path being importable. New code should import from
+``matgl.models._megnet_dgl`` (or use the public re-export from
+``matgl.models``).
 """
 
 from __future__ import annotations
 
-import logging
-from typing import TYPE_CHECKING
+from matgl.models._megnet_dgl import MEGNet
 
-import torch
-from dgl.nn import Set2Set
-from torch import nn
-
-from matgl.config import DEFAULT_ELEMENTS
-from matgl.graph._compute_dgl import compute_pair_vector_and_distance
-from matgl.layers import MLP, ActivationFunction, BondExpansion, EdgeSet2Set, EmbeddingBlock, MEGNetBlock
-
-from ._core import MatGLModel
-
-if TYPE_CHECKING:
-    import dgl
-
-    from matgl.graph._converters_dgl import GraphConverter
-
-logger = logging.getLogger(__file__)
-
-
-class MEGNet(MatGLModel):
-    """DGL implementation of MEGNet."""
-
-    __version__ = 1
-
-    def __init__(
-        self,
-        dim_node_embedding: int = 16,
-        dim_edge_embedding: int = 100,
-        dim_state_embedding: int = 2,
-        ntypes_state: int | None = None,
-        nblocks: int = 3,
-        hidden_layer_sizes_input: tuple[int, ...] = (64, 32),
-        hidden_layer_sizes_conv: tuple[int, ...] = (64, 64, 32),
-        hidden_layer_sizes_output: tuple[int, ...] = (32, 16),
-        nlayers_set2set: int = 1,
-        niters_set2set: int = 2,
-        activation_type: str = "softplus2",
-        is_classification: bool = False,
-        include_state: bool = True,
-        dropout: float = 0.0,
-        element_types: tuple[str, ...] = DEFAULT_ELEMENTS,
-        bond_expansion: BondExpansion | None = None,
-        cutoff: float = 4.0,
-        gauss_width: float = 0.5,
-        **kwargs,
-    ):
-        """Useful defaults for all arguments have been specified based on MEGNet formation energy model.
-
-        Args:
-            dim_node_embedding: Dimension of node embedding.
-            dim_edge_embedding: Dimension of edge embedding.
-            dim_state_embedding: Dimension of state embedding.
-            ntypes_state: Number of state types.
-            nblocks: Number of blocks.
-            hidden_layer_sizes_input: Architecture of dense layers before the graph convolution
-            hidden_layer_sizes_conv: Architecture of dense layers for message and update functions
-            nlayers_set2set: Number of layers in Set2Set layer
-            niters_set2set: Number of iterations in Set2Set layer
-            hidden_layer_sizes_output: Architecture of dense layers for concatenated features after graph convolution
-            activation_type: Activation used for non-linearity
-            is_classification: Whether this is classification task or not
-            layer_node_embedding: Architecture of embedding layer for node attributes
-            layer_edge_embedding: Architecture of embedding layer for edge attributes
-            layer_state_embedding: Architecture of embedding layer for state attributes
-            include_state: Whether the state embedding is included
-            dropout: Randomly zeroes some elements in the input tensor with given probability (0 < x < 1) according to
-                a Bernoulli distribution. Defaults to 0, i.e., no dropout.
-            element_types: Elements included in the training set
-            bond_expansion: Gaussian expansion for edge attributes
-            cutoff: cutoff for forming bonds
-            gauss_width: width of Gaussian function for bond expansion
-            **kwargs: For future flexibility. Not used at the moment.
-        """
-        super().__init__()
-
-        self.save_args(locals(), kwargs)
-
-        self.element_types = element_types or DEFAULT_ELEMENTS
-        self.cutoff = cutoff
-        self.bond_expansion = bond_expansion or BondExpansion(
-            rbf_type="Gaussian", initial=0.0, final=cutoff + 1.0, num_centers=dim_edge_embedding, width=gauss_width
-        )
-
-        node_dims = [dim_node_embedding, *hidden_layer_sizes_input]
-        edge_dims = [dim_edge_embedding, *hidden_layer_sizes_input]
-        state_dims = [dim_state_embedding, *hidden_layer_sizes_input]
-
-        try:
-            activation: nn.Module = ActivationFunction[activation_type].value()
-        except KeyError:
-            raise ValueError(
-                f"Invalid activation type, please try using one of {[af.name for af in ActivationFunction]}"
-            ) from None
-
-        self.embedding = EmbeddingBlock(
-            degree_rbf=dim_edge_embedding,
-            dim_node_embedding=dim_node_embedding,
-            ntypes_node=len(self.element_types),
-            ntypes_state=ntypes_state,
-            include_state=include_state,
-            dim_state_embedding=dim_state_embedding,
-            activation=activation,
-        )
-
-        self.edge_encoder = MLP(edge_dims, activation, activate_last=True)
-        self.node_encoder = MLP(node_dims, activation, activate_last=True)
-        self.state_encoder = MLP(state_dims, activation, activate_last=True)
-
-        dim_blocks_in = hidden_layer_sizes_input[-1]
-        dim_blocks_out = hidden_layer_sizes_conv[-1]
-        block_args = {
-            "conv_hiddens": hidden_layer_sizes_conv,
-            "dropout": dropout,
-            "act": activation,
-            "skip": True,
-        }
-        # first block
-        blocks = [MEGNetBlock(dims=[dim_blocks_in], **block_args)] + [  # type: ignore
-            MEGNetBlock(dims=[dim_blocks_out, *hidden_layer_sizes_input], **block_args)  # type: ignore
-            for _ in range(nblocks - 1)
-        ]
-
-        self.blocks = nn.ModuleList(blocks)
-
-        s2s_kwargs = {"n_iters": niters_set2set, "n_layers": nlayers_set2set}
-        self.edge_s2s = EdgeSet2Set(dim_blocks_out, **s2s_kwargs)
-        self.node_s2s = Set2Set(dim_blocks_out, **s2s_kwargs)
-
-        self.output_proj = MLP(
-            # S2S cats q_star to output producing double the dim
-            dims=[2 * 2 * dim_blocks_out + dim_blocks_out, *hidden_layer_sizes_output, 1],
-            activation=activation,
-            activate_last=False,
-        )
-
-        self.dropout = nn.Dropout(dropout) if dropout else None
-
-        self.is_classification = is_classification
-        self.include_state_embedding = include_state
-
-    def forward(self, g: dgl.DGLGraph, state_attr: torch.Tensor | None = None, **kwargs):
-        """Forward pass of MEGnet. Executes all blocks.
-
-        Args:
-            g (dgl.DGLGraph): DGL graphs
-            state_attr (torch.Tensor): State attributes
-            **kwargs: For future flexibility. Not used at the moment.
-
-        Returns:
-            Prediction
-        """
-        node_attr = g.ndata["node_type"]
-        bond_vec, bond_dist = compute_pair_vector_and_distance(g)
-        g.edata["bond_vec"] = bond_vec
-        g.edata["bond_dist"] = bond_dist
-        edge_attr = self.bond_expansion(g.edata["bond_dist"])
-        node_feat, edge_feat, state_feat = self.embedding(node_attr, edge_attr, state_attr)
-        edge_feat = self.edge_encoder(edge_feat)
-        node_feat = self.node_encoder(node_feat)
-        state_feat = self.state_encoder(state_feat)
-
-        for block in self.blocks:
-            output = block(g, edge_feat, node_feat, state_feat)
-            edge_feat, node_feat, state_feat = output
-
-        node_vec = self.node_s2s(g, node_feat)
-        edge_vec = self.edge_s2s(g, edge_feat)
-
-        node_vec = torch.squeeze(node_vec)
-        edge_vec = torch.squeeze(edge_vec)
-        state_feat = torch.squeeze(state_feat)
-
-        vec = torch.hstack([node_vec, edge_vec, state_feat])
-
-        if self.dropout:
-            vec = self.dropout(vec)  # pylint: disable=E1102
-
-        output = self.output_proj(vec)
-        if self.is_classification:
-            output = torch.sigmoid(output)
-
-        return torch.squeeze(output)
-
-    def predict_structure(
-        self,
-        structure,
-        state_attr: torch.Tensor | None = None,
-        graph_converter: GraphConverter | None = None,
-    ):
-        """Convenience method to directly predict property from structure.
-
-        Args:
-            structure: An input crystal/molecule.
-            state_attr (torch.tensor): Graph attributes
-            graph_converter: Object that implements a get_graph_from_structure.
-
-        Returns:
-            output (torch.tensor): output property
-        """
-        if graph_converter is None:
-            from matgl.ext._pymatgen_dgl import Structure2Graph
-
-            graph_converter = Structure2Graph(element_types=self.element_types, cutoff=self.cutoff)
-        g, lat, state_attr_default = graph_converter.get_graph(structure)
-        g.edata["pbc_offshift"] = torch.matmul(g.edata["pbc_offset"], lat[0])
-        g.ndata["pos"] = g.ndata["frac_coords"] @ lat[0]
-        if state_attr is None:
-            state_attr = torch.tensor(state_attr_default)
-        _, bond_dist = compute_pair_vector_and_distance(g)
-        g.edata["edge_attr"] = self.bond_expansion(bond_dist)
-        return self(g=g, state_attr=state_attr).detach()
+__all__ = ["MEGNet"]
